remove comment before create package
This makes it compatible with the new
ut_trigger_annotation_parsing trigger, which currently does
not expect to have leading comments. This is possible with low
level interfaces such as JDBC only. Tools such as
SQL*Plus, SQLcl, SQL Developer, etc. do not pass the comment to
the SQL engine.