add alternative tree-table-view based on #65 to the RunnerPanel
Columns in table and tree-table are shown in the same order.
This means the warning and info indicators are shown now before
the execution time column.

All features that were available for the table-view are also available
in the tree-table-view with one exception: It is not possible to sort a
tree-table. Therefore sorting is disabled for the tree-table-view.
Beside that, all  features that are available in the context-menu
work. Selection multiple nodes is supported as well.